S28-4920 split out ffmpeg, and program to editing service interface (2nd PR) (#1420)
