--- src/main/java/uk/gov/hmcts/reform/preapi/media/edit/IEditingService.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.reform.preapi.media.edit;
-
-import uk.gov.hmcts.reform.preapi.entities.EditRequest;
-
-import java.util.UUID;
-
-@SuppressWarnings("PMD.ImplicitFunctionalInterface")
-public interface IEditingService {
-    void performEdit(UUID newRecordingId, EditRequest request);
-}
